chore: remove deprecated errors mapping
