+++ ts/features/it-wallet/screens/issuing/pid/ItwPidAuthInfoScreen.tsx
@@ -24,8 +24,10 @@ import { itwWiaRequest } from "../../../store/actions/itwWiaActions";
 import ItwContinueView from "../../../components/ItwContinueView";
 import ItwLoadingSpinnerOverlay from "../../../components/ItwLoadingSpinnerOverlay";
 import ItwKoView from "../../../components/ItwKoView";
-import { getItwGenericMappedError } from "../../../utils/errors/itwErrorsMapping";
-import { ItWalletError } from "../../../utils/errors/itwErrors";
+import {
+  getItwGenericMappedError,
+  ItWalletError
+} from "../../../utils/itwErrorsUtils";
 
 /**
  * Renders the screen which displays the information about the authentication process to obtain a Wallet Instance.
