+++ ts/features/it-wallet/store/actions/itwWiaActions.ts
@@ -1,5 +1,5 @@
 import { ActionType, createAsyncAction } from "typesafe-actions";
-import { ItWalletError } from "../../utils/errors/itwErrors";
+import { ItWalletError } from "../../utils/itwErrorsUtils";
 
 /**
  * Start ITW wallet instance attestation request.
