+++ ts/features/it-wallet/store/actions/new/itwPresentationActions.ts
@@ -1,5 +1,5 @@
 import { ActionType, createAsyncAction } from "typesafe-actions";
-import { ItWalletError } from "../../../utils/errors/itwErrors";
+import { ItWalletError } from "../../../utils/itwErrorsUtils";
 
 /**
  * Async action which starts the presentation checks.
