+++ ts/features/it-wallet/store/reducers/itwRpPresentationReducer.ts
@@ -2,7 +2,7 @@ import { getType } from "typesafe-actions";
 import * as pot from "@pagopa/ts-commons/lib/pot";
 import * as O from "fp-ts/lib/Option";
 import { Action } from "../../../../store/actions/types";
-import { ItWalletError } from "../../utils/errors/itwErrors";
+import { ItWalletError } from "../../utils/itwErrorsUtils";
 import { GlobalState } from "../../../../store/reducers/types";
 import { itwRpPresentation } from "../actions/itwRpActions";
