+++ ts/features/it-wallet/store/reducers/new/itwPresentationReducer.ts
@@ -5,7 +5,7 @@ import * as pot from "@pagopa/ts-commons/lib/pot";
 import { getType } from "typesafe-actions";
 import { Action } from "../../../../../store/actions/types";
 import { GlobalState } from "../../../../../store/reducers/types";
-import { ItWalletError } from "../../../utils/errors/itwErrors";
+import { ItWalletError } from "../../../utils/itwErrorsUtils";
 import {
   itwPresentation,
   itwPresentationChecks
